pyproject.toml
[project]
name = "excelqa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.3.0",
    "langchain-google-genai>=1.0.0",
    "langchain>=0.1.0",
    "numpy>=1.24.0",
    "openpyxl>=3.1.0",
    "pandas>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "langchain-experimental>=0.0.47",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.34.0",
    "python-multipart>=0.0.6",
    "langchain-core>=0.1.0",
]
